GroupsController class + changes in UsersController relation many-to-many to models classes added

+++ app/Http/Controllers/UsersController.php
@@ -48,6 +48,8 @@ class UsersController extends Controller
             $user->date_of_birth= $request->input('date_of_birth');
 
             $user -> save();
+
+            $user -> groups() -> sync($request->input('group_id'));
             return response()->json($user);
 
         }
@@ -62,7 +64,9 @@ class UsersController extends Controller
     public function show($id)
     {
         $user = MyUser::find($id);
-        return response()->json($user);
+        $groups = $user->groups()->get();
+       // return response()->json($user);
+        return array('user' => $user, 'groups' => $groups);
     }
 
 
@@ -95,6 +99,7 @@ class UsersController extends Controller
             $user->date_of_birth= $request->input('date_of_birth');
 
             $user -> save();
+            $user -> groups() -> syn($request->input('group_id'));
             return response()->json($user);
 
         }
@@ -104,14 +109,37 @@ class UsersController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return array
      */
     public function destroy($id)
     {
         $user=MyUser::find($id);
-        $user->delete();
+        if($user!=null)
+        {
 
-        $response = array('response' => 'User deleted', 'success'=> true);
+            $user->delete();
+            $response = array('response' => 'User deleted', 'success'=> true);
+        }
+        else
+        {
+            $response= array('response' => "User doesn't exist", 'success'=> false);
+        }
+
         return $response;
+    }
+
+    public function userGroups($id)
+    {
+        $user= MyUser::find($id);
+        if($user)
+        {
+            return $user->groups()->get();
+        }
+        else
+        {
+            return array('response' => "User doesn't exist", 'success'=> false);
+        }
+
+
     }
 }
